separate model and methods
