+++ lib/features/basic_giving_flow/pages/donation.dart
@@ -3,7 +3,8 @@ import 'package:flutter/gestures.dart';
 import 'package:givt_mobile_apps/core/widgets/buttons/generic_button.dart';
 import 'package:givt_mobile_apps/core/templates/donation_template.dart';
 import 'package:givt_mobile_apps/models/html.dart';
-import 'package:givt_mobile_apps/models/localStorage.dart';
+import 'package:givt_mobile_apps/models/local_storage.dart';
+import 'package:givt_mobile_apps/services/local_storage_service.dart';
 import 'package:givt_mobile_apps/utils/locator.dart';
 import 'package:logger/logger.dart';
 import 'package:flutter/material.dart';
