+++ lib/features/permissions/controllers/location_permission.dart
@@ -1,4 +1,5 @@
-import 'package:givt_mobile_apps/models/localStorage.dart';
+import 'package:givt_mobile_apps/models/local_storage.dart';
+import 'package:givt_mobile_apps/services/local_storage_service.dart';
 
 import 'package:givt_mobile_apps/services/location_service.dart';
 import 'package:givt_mobile_apps/services/navigation_service.dart';
