+++ lib/features/startup/pages/startup.dart
@@ -3,7 +3,8 @@ import 'package:carousel_slider/carousel_slider.dart';
 import 'package:flutter/material.dart' hide ProgressIndicator;
 import 'package:flutter_native_splash/flutter_native_splash.dart';
 import 'package:flutter_svg/svg.dart';
-import 'package:givt_mobile_apps/models/localStorage.dart';
+import 'package:givt_mobile_apps/models/local_storage.dart';
+import 'package:givt_mobile_apps/services/local_storage_service.dart';
 import 'package:givt_mobile_apps/utils/progress_indicator.dart';
 // ignore: import_of_legacy_library_into_null_safe
 import '../../../core/templates/logo_header_template.dart';
