+++ lib/utils/locator.dart
@@ -2,7 +2,7 @@ import 'package:get_it/get_it.dart';
 import 'package:givt_mobile_apps/services/camera_service.dart';
 import 'package:givt_mobile_apps/services/location_service.dart';
 import 'package:givt_mobile_apps/services/user_service.dart';
-import '../models/localStorage.dart';
+import '../services/local_storage_service.dart';
 import '../services/navigation_service.dart';
 
 GetIt locator = GetIt.instance;
